Extract entry transfer controller
Move transfer creation, editing, linking, and unlinking endpoints out of EntryController into EntryTransferController. Add EntryListingModelAssembler to share listing model assembly for HTMX refreshes without coupling controllers together. Move transfer-specific tests to EntryTransferControllerTest and keep EntryControllerTest focused on entry behavior.
