+++ src/test/java/dev/ccosta/aisha/web/entry/EntryControllerTest.java
@@ -2,9 +2,6 @@ package dev.ccosta.aisha.web.entry;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.ArgumentMatchers.anyBoolean;
-import static org.mockito.ArgumentMatchers.anyInt;
-import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
@@ -13,12 +10,9 @@ import dev.ccosta.aisha.application.category.CategoryOption;
 import dev.ccosta.aisha.application.category.CategoryService;
 import dev.ccosta.aisha.application.entry.EntryCategorySelection;
 import dev.ccosta.aisha.application.entry.categorization.EntryCategorySuggestionService;
-import dev.ccosta.aisha.application.entry.EntryNotFoundException;
 import dev.ccosta.aisha.application.entry.EntryService;
 import dev.ccosta.aisha.application.entry.transfer.EntryTransferService;
-import dev.ccosta.aisha.application.entry.importing.statement.EntryStatementImportService;
 import dev.ccosta.aisha.domain.entry.Entry;
-import dev.ccosta.aisha.domain.shared.PagedResult;
 import dev.ccosta.aisha.web.timefilter.DateFilterState;
 import java.math.BigDecimal;
 import java.time.LocalDate;
@@ -53,10 +47,7 @@ class EntryControllerTest {
     private EntryTransferService entryTransferService;
 
     @Mock
-    private EntryStatementImportService entryStatementImportService;
-
-    @Mock
-    private EntryImportJobCoordinator entryImportJobCoordinator;
+    private EntryListingModelAssembler listingModelAssembler;
 
     @Mock
     private MessageSource messageSource;
@@ -129,148 +120,8 @@ class EntryControllerTest {
     }
 
     @Test
-    void shouldLinkTransferFromListingSelectionAndReturnSuccessToast() {
-        ConcurrentModel model = new ConcurrentModel();
-        stubListingDependencies();
-        when(messageSource.getMessage(eq("entries.list.toast.transfer.link.success"), eq(null), any()))
-            .thenReturn("Transferência associada com sucesso.");
-
-        String view = entryController.linkTransferFromSelection(
-            1L,
-            List.of(1L, 2L),
-            baseDateFilter(),
-            null,
-            null,
-            null,
-            false,
-            2,
-            25,
-            htmxRequest(),
-            model
-        );
-
-        assertThat(view).isEqualTo("entries/list :: listing");
-        assertThat(model.getAttribute("toastMessage")).isEqualTo("Transferência associada com sucesso.");
-        assertThat(model.getAttribute("toastLevel")).isEqualTo("success");
-        verify(entryTransferService).linkExistingEntries(1L, 2L);
-    }
-
-    @Test
-    void shouldReturnErrorToastWhenNoCounterpartIsSelected() {
-        ConcurrentModel model = new ConcurrentModel();
-        stubListingDependencies();
-        when(messageSource.getMessage(eq("entries.list.toast.transfer.link.selection.single"), eq(null), any()))
-            .thenReturn("Selecione exatamente um outro lançamento para associar como transferência.");
-
-        String view = entryController.linkTransferFromSelection(
-            1L,
-            List.of(1L),
-            baseDateFilter(),
-            null,
-            null,
-            null,
-            false,
-            2,
-            25,
-            htmxRequest(),
-            model
-        );
-
-        assertThat(view).isEqualTo("entries/list :: listing");
-        assertThat(model.getAttribute("toastLevel")).isEqualTo("error");
-        assertThat(model.getAttribute("toastMessage"))
-            .isEqualTo("Selecione exatamente um outro lançamento para associar como transferência.");
-    }
-
-    @Test
-    void shouldReturnErrorToastWhenMoreThanOneCounterpartIsSelected() {
-        ConcurrentModel model = new ConcurrentModel();
-        stubListingDependencies();
-        when(messageSource.getMessage(eq("entries.list.toast.transfer.link.selection.single"), eq(null), any()))
-            .thenReturn("Selecione exatamente um outro lançamento para associar como transferência.");
-
-        String view = entryController.linkTransferFromSelection(
-            1L,
-            List.of(1L, 2L, 3L),
-            baseDateFilter(),
-            null,
-            null,
-            null,
-            false,
-            2,
-            25,
-            htmxRequest(),
-            model
-        );
-
-        assertThat(view).isEqualTo("entries/list :: listing");
-        assertThat(model.getAttribute("toastLevel")).isEqualTo("error");
-        assertThat(model.getAttribute("toastMessage"))
-            .isEqualTo("Selecione exatamente um outro lançamento para associar como transferência.");
-    }
-
-    @Test
-    void shouldReturnExplainedErrorToastWhenSelectedCounterpartIsIncompatible() {
-        ConcurrentModel model = new ConcurrentModel();
-        stubListingDependencies();
-        when(messageSource.getMessage(eq("entries.list.toast.transfer.link.incompatible"), any(), any()))
-            .thenReturn("O lançamento selecionado não pode ser associado: os lançamentos devem ter sinais opostos");
-        when(entryTransferService.linkExistingEntries(1L, 2L))
-            .thenThrow(new IllegalArgumentException("Transfer entries must have opposite signs"));
-
-        String view = entryController.linkTransferFromSelection(
-            1L,
-            List.of(1L, 2L),
-            baseDateFilter(),
-            null,
-            null,
-            null,
-            false,
-            2,
-            25,
-            htmxRequest(),
-            model
-        );
-
-        assertThat(view).isEqualTo("entries/list :: listing");
-        assertThat(model.getAttribute("toastLevel")).isEqualTo("error");
-        assertThat(model.getAttribute("toastMessage"))
-            .isEqualTo("O lançamento selecionado não pode ser associado: os lançamentos devem ter sinais opostos");
-    }
-
-    @Test
-    void shouldReturnExplainedErrorToastWhenSelectedCounterpartIsMissing() {
-        ConcurrentModel model = new ConcurrentModel();
-        stubListingDependencies();
-        when(messageSource.getMessage(eq("entries.list.toast.transfer.link.incompatible"), any(), any()))
-            .thenReturn("O lançamento selecionado não pode ser associado: o lançamento selecionado não foi encontrado");
-        when(entryTransferService.linkExistingEntries(1L, 2L))
-            .thenThrow(new EntryNotFoundException(2L));
-
-        String view = entryController.linkTransferFromSelection(
-            1L,
-            List.of(1L, 2L),
-            baseDateFilter(),
-            null,
-            null,
-            null,
-            false,
-            2,
-            25,
-            htmxRequest(),
-            model
-        );
-
-        assertThat(view).isEqualTo("entries/list :: listing");
-        assertThat(model.getAttribute("toastLevel")).isEqualTo("error");
-        assertThat(model.getAttribute("toastMessage"))
-            .isEqualTo("O lançamento selecionado não pode ser associado: o lançamento selecionado não foi encontrado");
-    }
-
-    @Test
     void shouldConfirmCategorySuggestionsInBulkAndRefreshListingForHtmx() {
         ConcurrentModel model = new ConcurrentModel();
-        stubListingDependencies();
 
         String view = entryController.bulkConfirmCategorySuggestions(
             List.of(1L, 2L),
@@ -287,13 +138,6 @@ class EntryControllerTest {
 
         assertThat(view).isEqualTo("entries/list :: listing");
         verify(entryService).bulkConfirmCategorySuggestions(List.of(1L, 2L));
-    }
-
-    private void stubListingDependencies() {
-        when(accountService.listVisibleForEntryFilter(any())).thenReturn(List.of());
-        when(categoryService.listHierarchyOptions()).thenReturn(List.of());
-        when(entryService.listMostRecentBySettlementDateBetweenAndFilters(any(), any(), any(), any(), any(), anyBoolean(), anyBoolean(), anyInt(), anyInt()))
-            .thenReturn(new PagedResult<>(List.of(), 0, 25, 0, 0));
     }
 
     private DateFilterState baseDateFilter() {
